+++ lib/flight_job/question_generators/file_listing.rb
@@ -26,7 +26,7 @@
 #==============================================================================
 
 module FlightJob
-  module OptionGenerators
+  module QuestionGenerators
 
     # Generate a list of options based on globbing a directory or directories.
     #
